requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pymongo==3.12.0
requests==2.26.0
beautifulsoup4==4.10.0
python-multipart==0.0.5
joblib==1.1.0
scikit-learn==1.0.1
numpy==1.21.2
pydantic==1.8.2
jinja2==3.0.1
# Computer Vision
opencv-python-headless==4.5.5.64
pillow==9.0.1
python-whois==0.7.3
